Добавлен компонент CardList.js
